+++ test/Unit/VAST/VastEndScreenEventHandlerTest.ts
@@ -133,7 +133,7 @@ import { IPurchasingApi } from 'Purchasing/IPurchasing';
             vastEndScreenParameters = {
                 campaign: vastAdUnitParameters.campaign,
                 clientInfo: vastAdUnitParameters.clientInfo,
-                seatId: vastAdUnitParameters.campaign.getSeatId()
+                country: vastAdUnitParameters.coreConfig.getCountry()
             };
         });